+++ templates/layout/custom.php
@@ -11,20 +11,22 @@ $contentPosition = $options['structure'] == 'sidebar_left' ? 'right' : 'left';
 <div style="width: <?php echo $options['page_width']; ?>; margin-left: auto; margin-right: auto">
     <?php if($options['structure'] == 'sidebar_left') : ?>
         <div style="float: left; width: <?php echo $options['proportions']['sidebar'].'%';?>">
-            <?php do_action('jigoshop\sidebar'); ?>
+            <?php do_action('jigoshop\template\sidebar'); ?>
             <?php dynamic_sidebar('jigoshop_sidebar_'.$options['sidebar']); ?>
         </div>
     <?php endif; ?>
     <div style="width: <?php echo $contentSize; ?>; float: <?php echo $contentPosition; ?>">
         <div id="jigoshop_content" role="main" class="jigoshop">
-            <?php do_action('jigoshop\shop\content\before'); ?>
+            <?php /** @deprecated */ do_action('jigoshop\shop\content\before'); ?>
+            <?php do_action('jigoshop\template\shop\content\before'); ?>
             <?php echo $content; ?>
-            <?php do_action('jigoshop\shop\content\after'); ?>
+            <?php /** @deprecated */ do_action('jigoshop\shop\content\after'); ?>
+            <?php do_action('jigoshop\template\shop\content\after'); ?>
         </div>
     </div>
     <?php if($options['structure'] == 'sidebar_right') : ?>
         <div style="float: right; width: <?php echo $options['proportions']['sidebar'].'%';?>">
-            <?php do_action('jigoshop\sidebar'); ?>
+            <?php do_action('jigoshop\template\sidebar'); ?>
             <?php dynamic_sidebar('jigoshop_sidebar_'.$options['sidebar']); ?>
         </div>
     <?php endif; ?>