+++ templates/layout/twentyseventeen.php
@@ -7,14 +7,16 @@ get_header('shop');
 <div class="wrap">
     <div id="primary" class="content-area">
         <main id="main" class="site-main jigoshop" role="main">
-            <?php do_action('jigoshop\shop\content\before'); ?>
+            <?php /** @deprecated */ do_action('jigoshop\shop\content\before'); ?>
+            <?php do_action('jigoshop\template\shop\content\before'); ?>
             <div class="content">
                 <?php echo $content; ?>
             </div>
-            <?php do_action('jigoshop\shop\content\after'); ?>
+            <?php /** @deprecated */ do_action('jigoshop\shop\content\after'); ?>
+            <?php do_action('jigoshop\template\shop\content\after'); ?>
         </main>
     </div>
-    <?php do_action('jigoshop\sidebar'); ?>
+    <?php do_action('jigoshop\template\sidebar'); ?>
     <?php get_sidebar('shop'); // TODO: Remove on implementation of jigoshop\sidebar ?>
 </div>
 <?php get_footer('shop');