+++ templates/shop/checkout/pay.php
@@ -29,13 +29,15 @@ use Jigoshop\Helper\Render;
 			<th class="product-quantity"><?php _e('Quantity', 'jigoshop'); ?></th>
 			<th class="product-subtotal"><?php _e('Price', 'jigoshop'); ?></th>
 		</tr>
-		<?php do_action('jigoshop\checkout\table_head', $order); ?>
+		<?php /** @deprecated */ do_action('jigoshop\checkout\table_head', $order); ?>
+		<?php do_action('jigoshop\template\checkout\table_head', $order); ?>
 		</thead>
 		<tbody>
 		<?php foreach($order->getItems() as $key => $item): /** @var $item \Jigoshop\Entity\Order\Item */ ?>
 			<?php Render::output('shop/checkout/item/'.$item->getType(), array('cart' => $order, 'key' => $key, 'item' => $item, 'showWithTax' => $showWithTax, 'suffix' => $suffix)); ?>
 		<?php endforeach; ?>
-		<?php do_action('jigoshop\checkout\table_body', $order); ?>
+		<?php /** @deprecated */ do_action('jigoshop\checkout\table_body', $order); ?>
+		<?php do_action('jigoshop\template\checkout\table_body', $order); ?>
 		</tbody>
 		<tfoot>
 		<tr id="product-subtotal">
@@ -43,7 +45,10 @@ use Jigoshop\Helper\Render;
 			<th scope="row" colspan="4" class="text-right"><?php _e('Products subtotal', 'jigoshop'); ?></th>
 			<td><?php echo Product::formatPrice($productSubtotal); ?></td>
 		</tr>
-		</tfoot>
+        <?php /** @deprecated */ do_action('jigoshop\checkout\table_foot', $order); ?>
+        <?php do_action('jigoshop\template\checkout\table_foot', $order); ?>
+
+        </tfoot>
 	</table>
 	<div id="cart-collaterals">
 		<div id="cart-totals" class="panel panel-primary pull-right">