+++ src/Jigoshop/Frontend/Page/Account/ChangePassword.php
@@ -34,7 +34,7 @@ class ChangePassword implements PageInterface
 
 		Styles::add('jigoshop.user.account', \JigoshopInit::getUrl().'/assets/css/user/account.css');
 		Styles::add('jigoshop.user.account.change_password', \JigoshopInit::getUrl().'/assets/css/user/account/change_password.css', array('jigoshop.user.account'));
-		$this->wp->doAction('jigoshop\account\assets', $wp);
+		$this->wp->doAction('jigoshop\account\change_password\assets', $wp);
 	}
 
 	public function action()