+++ src/Jigoshop/Frontend/Page/Account/EditAddress.php
@@ -41,7 +41,7 @@ class EditAddress implements PageInterface
 
 		Scripts::add('jigoshop.vendors.select2', \JigoshopInit::getUrl().'/assets/js/vendors/select2.js', array('jquery'));
 		Scripts::add('jigoshop.vendors.bs_tab_trans_tooltip_collapse', \JigoshopInit::getUrl().'/assets/js/vendors/bs_tab_trans_tooltip_collapse.js', array('jquery'));
-		$this->wp->doAction('jigoshop\account\assets', $wp);
+		$this->wp->doAction('jigoshop\account\edit_address\assets', $wp);
 	}
 
 	public function action()