+++ src/Jigoshop/Service/ProductService.php
@@ -16,7 +16,6 @@ use WPAL\Wordpress;
 
 /**
  * Product service.
- *
  * @package Jigoshop\Service
  * @author  Amadeusz Starzykiewicz
  */