+++ templates/admin/settings/layout/css.php
@@ -1,0 +1,11 @@
+<?php
+/**
+ *
+ */
+?>
+<div class="tooltip-inline-badge"></div>
+<div class="tooltip-inline-input">
+    <div class="col-xs-12">
+        <textarea class="form-control" style="width: 100%; height: 100px;" name="<?php echo $name;?>"><?php echo $value; ?></textarea>
+    </div>
+</div>