+++ templates/shop.php
@@ -13,7 +13,7 @@ use Jigoshop\Helper\Render;
 <?php if(is_search()): ?>
 	<h1 class="page-title"><?php _e('Search Results:', 'jigoshop'); ?> &ldquo;<?php the_search_query(); ?>&rdquo; <?php if (get_query_var('paged')) echo ' &mdash; Page '.get_query_var('paged'); ?></h1>
 <?php else: ?>
-	<?php echo apply_filters('jigoshop\shop\content\title', '<h1 class="page-title">'.$title.'</h1>'); ?>
+	<?php echo apply_filters('jigoshop\shop\content\title', '<h1 class="page-title">'.$title.'</h1>', $title); ?>
 <?php endif; ?>
 
 <?php Render::output('shop/messages', array('messages' => $messages)); ?>