+++ CHANGELOG.md
@@ -9,6 +9,7 @@
     * Fix: Fix fatal error on activation triggered by options helper.
     * Fix: Tax helper should return tax for Taxable products.
     * Fix: Properly interpret ignore meta queries option.
+    * Fix: Now coupons and emails are available via api.
 * 2.0.9 - 2017.02.13:
     * Fix: Support Twentyseventeen theme.
     * Fix: Improve product reviews.